Organized async API tests.
