+++ packages/mongo/package.js
@@ -94,12 +94,11 @@ Package.onTest(function (api) {
   // XXX test order dependency: the allow_tests "partial allow" test
   // fails if it is run before mongo_livedata_tests.
   api.addFiles('mongo_livedata_tests.js', ['client', 'server']);
-  api.addFiles('mongo_livedata_async_tests.js', ['client', 'server']);
   api.addFiles('upsert_compatibility_test.js', 'server');
   api.addFiles('allow_tests.js', ['client', 'server']);
   api.addFiles('collection_tests.js', ['client', 'server']);
+  api.addFiles('collection_async_tests.js', ['client', 'server']);
   api.addFiles('observe_changes_tests.js', ['client', 'server']);
   api.addFiles('oplog_tests.js', 'server');
   api.addFiles('doc_fetcher_tests.js', 'server');
-  api.addFiles('collection_async_tests.js', 'server');
 });
